Made a slight change to auctioning so that minbid is the price of property
Reason behind it: if the min bid was set at 0, allowing people to get
the property for less than what it costs, we might as well get rid of
the buy property function because everyone will want to auction (which
will be annoying because it's 10+ seconds per each auction). So I'm
doing this to deter people from auctioning all the time.